refactor: reorganize modules and centralize utilities
