+++ Makefile
@@ -6,10 +6,9 @@
 #    By: dimendon <dimendon@student.hive.fi>        +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2025/06/04 13:42:42 by dimendon          #+#    #+#              #
-#    Updated: 2025/08/25 10:01:35 by dimendon         ###   ########.fr        #
+#    Updated: 2025/08/26 by dimendon         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
-
 
 .MAKEFLAGS: --no-print-directory
 
@@ -26,34 +25,39 @@ OBJ_PATH    = obj/
 LIBFT_DIR   = src/libft
 LIBFT_LIB   = $(LIBFT_DIR)/libft.a
 
-SRCS =	        src/builtins/custom_export/export_check.c \
-				src/builtins/custom_export/export_env.c \
-				src/builtins/custom_export/export_print.c \
-				src/builtins/custom_export/export_utils.c \
-				src/builtins/custom_cd.c \
-		        src/builtins/custom_echo.c \
-				src/builtins/custom_env.c \
-                src/builtins/custom_exit.c \
-                src/builtins/custom_pwd.c \
-                src/builtins/custom_unset.c \
-				src/env/env_lookup.c \
-				src/env/env_utils.c \
-				src/misc/utils.c \
-				src/parsing/expander_utils.c \
-				src/parsing/expander.c \
-				src/parsing/redir_split.c \
-				src/parsing/split_pipes.c \
-				src/parsing/tokenize.c \
-				src/parsing/word_split.c \
-                               src/piping/pipeline.c \
-                               src/piping/redirections.c \
-                               src/piping/pipeline_utils.c \
-                               src/piping/redirections_utils.c \
-                               src/cleanup.c \
-                       src/command_processor.c \
-                               src/command_utils.c \
-               src/command_execution.c \
-               src/main.c
+SRCS = \
+    src/builtins/custom_export/export_check.c \
+    src/builtins/custom_export/export_env.c \
+    src/builtins/custom_export/export_print.c \
+    src/builtins/custom_export/export_utils.c \
+    src/builtins/custom_cd.c \
+    src/builtins/custom_echo.c \
+    src/builtins/custom_env.c \
+    src/builtins/custom_exit.c \
+    src/builtins/custom_pwd.c \
+    src/builtins/custom_unset.c \
+    src/env/env_lookup.c \
+    src/env/env_utils.c \
+    src/utils/array_utils.c \
+    src/utils/cleanup.c \
+    src/utils/error.c \
+    src/parsing/expansion/expander_utils.c \
+    src/parsing/expansion/expander.c \
+    src/parsing/expansion/word_split.c \
+    src/parsing/redirection/redir_split.c \
+    src/parsing/tokenization/split_pipes.c \
+    src/parsing/tokenization/tokenize.c \
+    src/piping/pipeline.c \
+    src/piping/redirections.c \
+    src/piping/pipeline_setup.c \
+    src/piping/pipeline_cmd.c \
+    src/piping/pipeline_helpers.c \
+    src/piping/redirections_utils.c \
+    src/piping/redirection_files.c \
+    src/execution/command_processor.c \
+    src/execution/command_utils.c \
+    src/execution/command_execution.c \
+    src/main.c
 
 OBJS = $(SRCS:src/%.c=$(OBJ_PATH)%.o)
 HEADERS = includes/minishell.h
@@ -78,7 +82,6 @@ $(TARGET): $(OBJS)
 $(LIBFT_LIB):
 	@$(MAKE) --no-print-directory -C $(LIBFT_DIR) build
 
-# Run with Valgrind
 run: $(TARGET)
 	@echo "$(YELLOW)Running with Valgrind...$(RESET)"
 	$(VALGRIND) ./$(TARGET)
