+++ includes/minishell.h
@@ -6,7 +6,7 @@
 /*   By: dimendon <dimendon@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/04 13:40:36 by dimendon          #+#    #+#             */
-/*   Updated: 2025/08/25 13:21:56 by dimendon         ###   ########.fr       */
+/*   Updated: 2025/08/26 by dimendon         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,39 +14,38 @@
 # define MINISHELL_H
 
 // Standard C Library
-# include <stdio.h>    // printf, perror
-# include <stdlib.h>   // malloc, free, exit, getenv
-# include <string.h>   // strerror
+# include <stdio.h>
+# include <stdlib.h>
+# include <string.h>
 
 // POSIX System Calls
-# include <unistd.h>   // write, access, read, close, fork, getcwd, chdir, execve, dup, dup2, pipe, isatty, ttyname, ttyslot
-# include <sys/wait.h> // wait, waitpid, wait3, wait4
-# include <sys/stat.h> // stat, lstat, fstat
-# include <fcntl.h>    // open
-# include <sys/ioctl.h>// ioctl
-# include <signal.h>   // signal, sigaction, sigemptyset, sigaddset, kill
+# include <unistd.h>
+# include <sys/wait.h>
+# include <sys/stat.h>
+# include <fcntl.h>
+# include <sys/ioctl.h>
+# include <signal.h>
 
 // Directory operations
-# include <dirent.h>   // opendir, readdir, closedir
+# include <dirent.h>
 
 // Termcap Library
-# include <termcap.h>  // tgetent, tgetflag, tgetnum, tgetstr, tgoto, tputs
+# include <termcap.h>
 
 // Terminal attributes
-# include <termios.h>  // tcsetattr, tcgetattr
+# include <termios.h>
 
 // GNU Readline Library
-# include <readline/readline.h> // readline, add_history
-# include <readline/history.h>  // rl_clear_history, rl_on_new_line, rl_replace_line, rl_redisplay
+# include <readline/readline.h>
+# include <readline/history.h>
 
 extern int  g_exit_code;
 
-// Tokenizer
 typedef struct s_token
 {
-    char    *str;       // actual token text
-    int     quoted;     // 0 = unquoted, 1 = single, 2 = double
-    int     type;       // e.g. WORD, REDIR_IN, REDIR_OUT, HEREDOC, PIPE
+    char    *str;
+    int     quoted;
+    int     type;
 }   t_token;
 
 typedef struct s_command
@@ -58,15 +57,14 @@ typedef struct s_command
     int     has_heredoc;
 }   t_command;
 
-// Pipeline context
 typedef struct s_pipeline
 {
     char **envp;
-	char **segments;
-	pid_t *pids;
-	int nbr_segments;
-	t_command *cmds;
-}	t_pipeline;
+    char **segments;
+    pid_t *pids;
+    int nbr_segments;
+    t_command *cmds;
+}       t_pipeline;
 
 typedef struct s_pipe_context
 {
@@ -89,58 +87,58 @@ void        init_export_index(int *index, int size);
 int         is_valid_name(const char *name);
 int         update_or_add_env(char ***env, char *arg);
 
-// ==================== COMMAND PROCESSOR ====================
-void        process_command(char ***envp, char *line);
+// ==================== EXECUTION ====================
 int         run_builtin(char ***envp, t_token **cmd);
-
-// ==================== COMMAND UTILS ====================
+void        process_command(char ***envp, char *line);
 int         is_folder(char *arg);
 t_token     **prepare_command(char *segment, int *in_fd, int *out_fd, char ***envp);
 void        setup_redirections(int in_fd, int out_fd, int *save_in, int *save_out);
 void        restore_redirections(int save_in, int save_out);
 short int   is_builtin(const char *cmd);
-
-// ==================== COMMAND EXECUTION ====================
 int         execute_command(char *path, t_token **tokens, char **envp);
+char        **prepare_argv_from_tokens(t_token **tokens);
 
-// ==================== ENV LOOKUP ====================
+// ==================== ENVIRONMENT ====================
 char        *get_env_value(char **envp, const char *name);
 char        *get_path(char **envp, char **cmd);
-
-// ==================== ENV UTILS ====================
 char        **copy_envp(char **envp);
 int         env_size(char **env);
 char        **env_realloc_add(char **env);
 int         env_add(char ***env_ptr, const char *new_var);
 
 // ==================== PIPING ====================
+void        execute_pipeline(char **envp, char **segments);
 void        execute_cmd(char **envp, t_token **cmd);
 void        close_pipe(int *fd);
 void        parent_cleanup(int *in_fd, int *fd, int i, int num);
 void        wait_for_all(pid_t *pids, int count);
-void        execute_pipeline(char **envp, char **segments);
 t_token     **handle_redirections(t_token **cmd, char **envp, int *in_fd, int *out_fd);
 int         handle_heredoc(const char *delim, int quoted, char **envp, int *in_fd);
 void        pipeline_loop(t_pipeline *pipeline);
+int         open_infile(char *file, int *in_fd);
+int         open_outfile(char *file, int *out_fd);
+int         open_appendfile(char *file, int *out_fd);
 
 // ==================== PARSING ====================
+// Tokenization
+t_token     *new_token(const char *str, int quoted, int type);
+t_token     **tokenize_command(char const *s, char c, char **envp);
+char        **split_pipes(const char *line);
+t_token     **split_redirs(t_token **arr);
+
+// Expansion
 void        remove_quotes(char *str);
 char        *append_literal(char *result, char *str, int start, int i);
 char        *expand_var(char *str, int *var_len);
 char        *append_expanded_var(char *result, char *str, int *i, char **envp);
 char        *build_expanded_str(char *str, char **envp);
 t_token     **split_expanded_tokens(t_token **arr);
-t_token     **split_redirs(t_token **arr);
-t_token     **tokenize_command(char const *s, char c, char **envp);
-char        **split_pipes(const char *line);
-t_token     *new_token(const char *str, int quoted, int type);
-void        free_tokens(t_token **arr);
-char        **prepare_argv_from_tokens(t_token **tokens);
 
-// ==================== MISC ====================
+// ==================== UTILS ====================
 void        free_cmd(char **cmd);
 void        free_tokens(t_token **arr);
-short int   is_builtin(const char *cmd);
 int         count_strings(char **arr);
+void        error_msg(const char *prefix, const char *msg);
+void        error_perror(const char *context);
 
 #endif
