+++ src/parsing/expansion/expander_utils.c
@@ -1,4 +1,4 @@
-#include "../libft/libft.h"
+#include "../../libft/libft.h"
 #include "minishell.h"
 
 void remove_quotes(char *str)
