+++ src/parsing/redirection/redir_split.c
@@ -1,4 +1,4 @@
-#include "../libft/libft.h"
+#include "../../libft/libft.h"
 #include "minishell.h"
 
 static int part_count(t_token *tok)
