+++ src/parsing/tokenization/split_pipes.c
@@ -1,4 +1,4 @@
-#include "../libft/libft.h"
+#include "../../libft/libft.h"
 #include "minishell.h"
 
 static size_t	count_segments(const char *line)
